Move system property / environment property configuration to belgif-rest-problem-java-ee
