+++ belgif-rest-problem-java-ee/src/test/java/io/github/belgif/rest/problem/jaxrs/i18n/I18NConfiguratorTest.java
@@ -9,6 +9,8 @@ import javax.servlet.ServletContextEvent;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.junitpioneer.jupiter.SetEnvironmentVariable;
+import org.junitpioneer.jupiter.SetSystemProperty;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
@@ -48,6 +50,50 @@ class I18NConfiguratorTest {
         when(servletContext.getInitParameter(I18N.I18N_FLAG)).thenReturn(null);
         configurator.contextInitialized(new ServletContextEvent(servletContext));
         assertThat(I18N.isEnabled()).isTrue();
+    }
+
+    @Test
+    @SetSystemProperty(key = I18N.I18N_FLAG, value = "false")
+    void disabledViaSystemProperty() {
+        configurator.contextInitialized(new ServletContextEvent(servletContext));
+        assertThat(I18N.isEnabled()).isFalse();
+        verifyNoInteractions(servletContext);
+    }
+
+    @Test
+    @SetSystemProperty(key = I18N.I18N_FLAG, value = "true")
+    void enabledViaSystemProperties() {
+        I18N.setEnabled(false);
+        configurator.contextInitialized(new ServletContextEvent(servletContext));
+        assertThat(I18N.isEnabled()).isTrue();
+        verifyNoInteractions(servletContext);
+    }
+
+    @Test
+    @SetEnvironmentVariable(key = I18N.I18N_FLAG, value = "false")
+    void disabledViaEnvironmentVariable() {
+        configurator.contextInitialized(new ServletContextEvent(servletContext));
+        assertThat(I18N.isEnabled()).isFalse();
+        verifyNoInteractions(servletContext);
+    }
+
+    @Test
+    @SetEnvironmentVariable(key = I18N.I18N_FLAG, value = "true")
+    void enabledViaEnvironmentVariable() {
+        I18N.setEnabled(false);
+        configurator.contextInitialized(new ServletContextEvent(servletContext));
+        assertThat(I18N.isEnabled()).isTrue();
+        verifyNoInteractions(servletContext);
+    }
+
+    @Test
+    @SetSystemProperty(key = I18N.I18N_FLAG, value = "true")
+    @SetEnvironmentVariable(key = I18N.I18N_FLAG, value = "false")
+    void systemPropertyHasPrecedenceOverEnvironmentVariable() {
+        I18N.setEnabled(false);
+        configurator.contextInitialized(new ServletContextEvent(servletContext));
+        assertThat(I18N.isEnabled()).isTrue();
+        verifyNoInteractions(servletContext);
     }
 
 }
