+++ src/main/asciidoc/index.adoc
@@ -507,7 +507,7 @@ io.github.belgif.rest.problem.i18n=false
 
 On Spring Boot, this can be configured via any of the standard https://docs.spring.io/spring-boot/reference/features/external-config.html[externalized configuration] mechanisms.
 
-On Jakarta EE, this can be configured as a web.xml context-param, system property or environment variable (in that order of precedence).
+On Jakarta EE, this can be configured as a system property, environment variable, or web.xml context-param (in that order of precedence).
 
 [[implementation-details]]
 == Implementation details
